backend/requirements.txt
streamlit
Pillow
opencv-python
tensorflow 
os
matplotlib
numpy
